Execute the docker binary to get the version
The Go library moves independently from the docker binary.
It can result of the library being too ahead from the server.

I run `docker version` and parse the output to extract what's the API
the server accepts.
Then use stick it into the environment: DOCKER_API_VERSION if it is not
already set.